refactor: consolidate turn test support modules (#3842)
Unify world-market, research, economy-preview, and naval test helpers under
test/support/ and test/naval/, removing four parallel world-market support
files and the stale research fixture fork.
