+++ packages/colonizethis_turn/test/research_phase_funding_pacing_test.dart
@@ -3,7 +3,7 @@ import 'package:colonizethis_data/colonizethis_data.dart';
 import 'package:colonizethis_logic/colonizethis_logic.dart';
 import 'package:colonizethis_models/colonizethis_models.dart';
 
-import 'research_phase_test_support.dart';
+import 'support/research_phase_test_support.dart';
 
 void main() {
   group('Research pacing at Medium funding (cost rebalance, Refs #3512)', () {
